# pkg/model/v1/resources/cpu.go
/*
 * Copyright (c) 2023 Kamillaova. All rights reserved.
 *
 * This software is licensed under the WebK8S End-User License Agreement (EULA).
 * A copy of the EULA is included in the repository's LICENSE file.
 *
 * For non-commercial usage only. Modifications must be published under the same license.
 */

package resources

type CPU struct {
	Model string `json:"name"`
	Cores uint32 `json:"cores"`
}